se agrego el obtener jurados por tema

+++ database/functions/functions_jurado.sql
@@ -609,12 +609,11 @@ $function$
 
 CREATE OR REPLACE FUNCTION obtener_jurados_por_tema(p_tema_id integer)
 RETURNS TABLE(
-    conteo integer
+    conteo bigint
 ) AS $$
 BEGIN
     RETURN QUERY
-    SELECT 
-        COUNT(*) as conteo
+    SELECT COUNT(*) as conteo
     FROM usuario_tema ut
     WHERE rol_id IN (1, 2, 5)
     AND ut.activo = true
